fix for unable to find rendering pipeline method
As Chris expected, Google broke this regex in MeetingsUIProxyHandler to find the rendering pipeline method again.  The only real hitch was the minified function name now had numbers in it as well.  May need to update other parts of the regex to allow numbers as well if it breaks again, but this is enough for now.

https://github.com/Fugiman/google-meet-grid-view/issues/225#issuecomment-656872533